requirements.txt
python-telegram-bot==13.15
Pillow
opencv-python
matplotlib
numpy
requests
flask